Worked on ChildNode function
I have no clue if this is what ChildNode is supposed to do so I gave it my best shot.  Right now I have it checking the state of Node n and returning the node to the left, right, above or below based on n's position(state) in the current state.  Feel free to edit or completely change it if you think its totally wrong.